+++ sgi-csp-service/src/main/java/org/crue/hercules/sgi/csp/dto/GrupoEnlaceInput.java
@@ -28,4 +28,6 @@ public class GrupoEnlaceInput implements Serializable {
   @NotNull
   private Long grupoId;
 
+  private Long tipoEnlaceId;
+
 }