+++ sgi-webapp/src/app/core/models/csp/grupo-enlace.ts
@@ -1,7 +1,9 @@
 import { IGrupo } from './grupo';
+import { ITipoEnlace } from './tipos-configuracion';
 
 export interface IGrupoEnlace {
   id: number;
   grupo: IGrupo;
   enlace: string;
+  tipoEnlace: ITipoEnlace;
 }