+++ sgi-webapp/src/app/core/services/csp/grupo-enlace/grupo-enlace-request.ts
@@ -2,4 +2,5 @@
 export interface IGrupoEnlaceRequest {
   grupoId: number;
   enlace: string;
+  tipoEnlaceId: number;
 }