+++ sgi-webapp/src/app/core/services/csp/grupo-enlace/grupo-enlace-response.ts
@@ -1,6 +1,6 @@
-
 export interface IGrupoEnlaceResponse {
   id: number;
   grupoId: number;
   enlace: string;
+  tipoEnlaceId: number;
 }